Add Blink to Mage logic
Logic:

If falling other than straight down and Slowfall is active and Blink is off CD then Blink as it doesn't break Slowfall.

If otherwise falling then toggle Slowfall on/off.

If running or swimming (actually moving) and Blink is off cooldown, use it. Additionally turn Slowfall off as it serves no purpose in this situation.

Combat Logic:

Changed Slowfall to Blink if useable as it's much more useful in a combat situation. Keep Slowfall as a backup in case Blink becomes spec-specific at some point.